+++ src/components/ApplicationForm/index.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import styled from 'styled-components'
-import { maxWidthMediaQueries } from './Common'
+import { maxWidthMediaQueries } from '../Common'
 
 const Container = styled.div`
   padding: 8vh 16vw;
